Add an extra argument 'procname' to ntt_pperl_new() which sets the default process name when no perl script is executing.  This process name appears in the output of `ps` and is also returned by getprogname(), making it appear in log messages.
Having to explicitely set the process name this way is slightly
unintuitive as one would expect (at least I would have) that the default
set up when the program starts would be retained unless explicitely
overridden.  In fact, perl's internals do override the process name; in
our case, it sets it to the very obscure string "-e".  To make matters
worse, there is no getproctitle() to get the original process name so we
cannot set the process name back to the original value within the
ntt_pperl library.

With this implementation, we let the caller specify what string to use as
the process name.  If getproctitle() were ever to be implemented, the
caller could just pass its output to ntt_pperl_new() to get the most
obvious behavior.  In the meantime, they can pass the output of
getprogname() to approximate the original process name.