Perl wisely protects descriptors 0 (stdin) and 1 (stdout) from being closed implicitely however does not extend this protection to the equally magical descriptor 2 (stderr).  As a result, if you override the perl STDERR handle, then perl will merrily close descriptor 2 allowing it to be reassigned to another file or socket.  The problem is that if you then call daemon(3), or any other routine which expects descriptors 0-2 to retain their special semantics, they will be operating on whatever has inherited descriptor #2 rather than (the expected) stderr stream.
This "feature" of perl's implementation bit me because I have a program
(routemond) which:
    a) overrides perl's STDERR file handle to log to a file.
    b) opens a socket.
    c) then calls daemon(3).

Perl's I/O library was closing file descriptor 2 in part a) which became
the first free descriptor so it as assigned to the socket opened in b).
Then daemon(3) redirected descriptors 0-2 to /dev/null in part c) causing
the socket to be closed.

Ideally, perl should extend the same protection to descriptor 2 that it
does for descriptors 0 and 1.  However, as a workaround, this delta
modifies ntt_pperl_override() to dup(2) descriptor 2 and restore it around
the call to Perl_do_close() to ensure that it does not get freed by perl.