Implement two new persistent perl-specific call lists, prologue and epilogue, and export routines to add code references to the lists. In persistent perl interpreters, BEGIN and END blocks are run whenever code is loaded or unloaded, respectively.  These new call lists allow routines to be specified to executed before and after each script run, respectively.
Prologue blocks are executed prior to running the main script.  If any
prologue block dies, the remaining prologue blocks are not executed nor is
the main script; instead, epilogue blocks are immediately executed.  This
is similar to the behavior of BEGIN blocks.

Epilogue blocks are executed after running the main script.  Epilogue
blocks can inspect $@ for errors raised by any previously-executed code
(e.g. prologue blocks, the main script, or other epilogue blocks).  If an
epilogue block dies, it merely changes the value of $@ and execution of
any remaining epilogue blocks continues.  This is similar to the behavior
of END blocks.

To syntax for adding code blocks to either of these two lists (from perl)
is:
	libpperl::prologue(sub { ... });
or
	libpperl::epilogue(sub { ... });

Prologue or epilogue blocks added by perl scripts will only be executed
when those scripts are run, however blocks added by perl modules are
executed for all scripts.